[F] Unify personPk type to int

+++ src/main/java/org/hydev/veracross/sdk/model/VeraLoginInfo.java
@@ -17,7 +17,7 @@ import lombok.experimental.Accessors;
 @Accessors(fluent = true)
 public class VeraLoginInfo
 {
-    long personPk;
+    int personPk;
     String username; // Email
     String securityRoles; // []
